# Conflicts:
#	src/main/java/com/goodee/corpdesk/employee/EmployeeController.java
#	src/main/java/com/goodee/corpdesk/employee/EmployeeService.java

+++ src/main/java/com/goodee/corpdesk/approval/dto/ApprovalDTO.java
@@ -15,7 +15,7 @@ public class ApprovalDTO {
 	private Long approvalId;
 	private String username;
 	private Integer departmentId;
-	private Long approvalFormId;
+	private Integer approvalFormId;
     private String approvalContent;
 	private Character status;
